feat: added server timing headers
